forgot password funtionality implemented

+++ controllers/user.js
@@ -1,58 +1,48 @@
-import { Sequelize, Op, where } from "sequelize";
+import { Sequelize, Op } from "sequelize";
+// import Redis from "redis";
 import User from "../models/user.js";
 import Post from "../models/posts.js";
-import { deletePostImage } from "../utils/deleteImages.js";
 import Follow from "../models/Follow.js";
 import Archive from "../models/Archive.js";
 import formatPostData from "../utils/dataFormater.js";
+import { deletePostImage } from "../utils/deleteImages.js";
+// import redisClient from "../utils/redisClient.js";
+
+const EXPIRATION = 3600;
 
 // Get user profile
 export const getUserProfile = async (req, res) => {
-  const id = req?.params?.id?.split(':')[1]||req.userId;
-  console.log(req.userId)
-console.log({'uaerId':req.params })
+  const id = req?.params?.id || req.userId;
+
   try {
+    // const cachedUserData = await redisClient.get(id);
+    // if (cachedUserData !== null) {
+    //   console.log('cach hit')
+    //   return res.status(200).json(JSON.parse(cachedUserData));
+    // }
+    // console.log('cach miss')
     const userInfo = await User.findOne({
-      where: {id},
-      attributes: { exclude: ['password'] }, // Exclude sensitive data
+      where: { id },
+      attributes: { exclude: ['password'] },
       include: [
-        {
-          model: User,
-          as: 'Followers',
-          through: { attributes: [] }, // Exclude join table attributes
-               attributes: []
-        },
-        {
-          model: User,
-          as: 'Following',
-          through: { attributes: [] }, // Exclude join table attributes
-          attributes: []
-        },{
-        model: Post,
-        as: 'SavedPosts',
-        through: { attributes: [] },
-      }
+        { model: User, as: 'Followers', through: { attributes: [] }, attributes: ['id'] },
+        { model: User, as: 'Following', through: { attributes: [] }, attributes: ['id'] },
+        { model: Post, as: 'SavedPosts', through: { attributes: [] } },
       ]
     });
 
     if (userInfo) {
-          const followersCount = await userInfo.countFollowers();
-      const followingCount = await userInfo.countFollowing();
-
-      const userProfile = {
-        ...userInfo.toJSON(),
-        followersCount,
-        followingCount,
-      };
-      res.status(200).json(userProfile); // Return user info
+      // await redisClient.setEx(id, EXPIRATION, JSON.stringify(userInfo));
+      return res.status(200).json(userInfo);
     } else {
-      res.status(404).json({ message: 'User not found' });
+      return res.status(404).json({ message: 'User not found' });
     }
   } catch (error) {
     console.error('Error fetching user profile:', error);
-    res.status(500).json({ message: 'An error occurred while fetching user profile' });
+    return res.status(500).json({ message: 'An error occurred while fetching user profile' });
   }
 };
+
 
 // Get posts by user ID
 export const getUserPostsById = async (req, res) => {
@@ -81,35 +71,42 @@ export const getUserPostsById = async (req, res) => {
 };
 
 // Follow a user
-export const FollowUser = async (req, res) => {
+export const FollowUser = async (req, res, next) => {
   const { followerId, followedId } = req.body;
+  console.log({ followerId, followedId });
 
   if (followerId === followedId) {
-    return res.status(400).json({ error: "You cannot follow yourself" });
+    return res.status(400).json({ status: "error", message: "You cannot follow yourself" });
   }
 
   try {
+    // Check if the follow relationship already exists
+    const existingFollow = await Follow.findOne({ where: { followerId, followedId } });
+
+    if (existingFollow) {
+      return res.status(200).json({ status: "success", message: "Already following this user" });
+    }
+
+    // Create a new follow relationship
     await Follow.create({ followerId, followedId });
+    await redisClient.del(followerId);
 
-    const userInfo = await User.findOne({
-      where: { id: req.userId },
-      attributes: { exclude: ['password'] },
-      include: [
-        { model: User, as: 'Followers', through: { attributes: [] } },
-        { model: User, as: 'Following', through: { attributes: [] } }
-      ]
-    });
-
-    res.status(201).json(userInfo); // Return updated user info
+    console.log('success')
+    res.status(201).json({ status: "success" });
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    if (error instanceof Sequelize.UniqueConstraintError) {
+      res.status(409).json({ status: "error", message: "You are already following this user" });
+    } else {
+      console.error("Error following user:", error);
+      next(error);
+    }
   }
 };
 
 // Unfollow a user
 export const Unfollow = async (req, res) => {
   const { followerId, followedId } = req.body;
-
+  console.log({ followerId, followedId })
   if (!followerId || !followedId) {
     return res.status(400).json({ error: "Missing followerId or followedId" });
   }
@@ -119,6 +116,7 @@ export const Unfollow = async (req, res) => {
 
     if (follow) {
       await follow.destroy();
+      await redisClient.del(followerId);
       res.status(200).json({ message: "Unfollowed successfully" });
     } else {
       res.status(400).json({ error: "Follow relationship does not exist" });
@@ -132,8 +130,7 @@ export const Unfollow = async (req, res) => {
 
 // Get followers of the current user
 export const getFollowers = async (req, res) => {
-  const userId = req?.params?.id?.split(':')[1]||req.userId;
-
+  const userId = req?.params?.userId || req.userId;
   try {
     const user = await User.findByPk(userId, {
       include: [{ model: User, as: 'Followers' }]
@@ -147,7 +144,7 @@ export const getFollowers = async (req, res) => {
 
 // Get users that the current user is following
 export const getFollowing = async (req, res) => {
-  const userId = req?.params?.id?.split(':')[1]||req.userId;
+  const userId = req?.params?.userId||req.userId;
 
   try {
     const user = await User.findByPk(userId, {
@@ -178,7 +175,7 @@ export const AddPostToArchive = async (req, res) => {
       PostId: postId,
       UserId: req.userId
     });
-
+          await redisClient.del(req.userId);
     res.status(200).json({ message: 'Post archived successfully', archived });
   } catch (error) {
     console.error('Error archiving post:', error);
@@ -197,6 +194,7 @@ try {
         });
   if (exist) {
     await exist.destroy();
+         await redisClient.del(UserId);
     res.status(200).json({message:'succesfully removes'})
   }
 } catch (error) {
@@ -251,9 +249,10 @@ export const EditUserProfile = async (req, res) => {
 
     if (data.removeImage && data.userImage && data.userImage !== 'null') {
       updatedData.userImage = null; // Remove user image
-      await deletePostImage([data.userImage]);
+      if (data.userFromOAuth === false) { await deletePostImage([data.userImage]) }
     }
 
+    delete data.userFromOAuth;
     const [_, updatedUser] = await User.update(updatedData, {
       where: { id: req.userId },
       attributes: { exclude: ['password'] },
